PAYEE-437: WhatDoYouWantToDoTileView.scala.html redirects to IncomeTaxHistoryView.scala.html
