+++ app/views/WhatDoYouWantToDoTileView.scala.html
@@ -69,7 +69,7 @@
 
                 <div class="flex-container grid-row">
                     @card(
-                        url = Some(s"/check-income-tax/historic-paye/${TaxYear().prev.year.toString}"),
+                        url = Some(s"/check-income-tax/income-tax-history"),
                         id = "earlierTaxYear",
                         heading = messages("earlier"),
                         bodyContent = Some(messages("check.tax.previous.years"))
