+++ conf/tai.routes
@@ -10,7 +10,7 @@ GET         /income-tax/no-info                                                 
 GET         /last-year-paye                                                            controllers.PayeControllerHistoric.lastYearPaye()
 GET         /historic-paye/:taxYear                                                    controllers.PayeControllerHistoric.payePage(taxYear: uk.gov.hmrc.tai.model.TaxYear)
 
-GET         /employment-history                                                         controllers.EmploymentHistoryController.onPageLoad()
+GET         /income-tax-history                                                        controllers.IncomeTaxHistoryController.onPageLoad()
 
 GET         /timeout                                                                   controllers.ServiceController.timeoutPage
 GET         /signout                                                                   controllers.ServiceController.serviceSignout
